pyproject.toml
[project]
name = "09-voiceagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "elevenlabs>=2.36.1",
    "fastapi>=0.133.1",
    "google-generativeai>=0.8.6",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.41.0",
    "websockets>=16.0",
]
